refactor: move auth functions to their own file
